+++ packages/instrumenter/testResources/instrumenter/optional-chains.ts
@@ -6,3 +6,5 @@ qux()?.map()
 const directiveRanges = comments?.map(tryParseTSDirective)
 
 const qux = quux(corge?.cov());
+
+input?.id!.toString();