+++ tasks/instrument-test-resources.js
@@ -12,11 +12,13 @@ const instrumenter = new Instrumenter({
 
 async function main() {
   await instrument({
-    './packages/mocha-runner/testResources/sample-project/MyMath.js': './packages/mocha-runner/testResources/sample-project-instrumented/MyMath.js'
+    './packages/mocha-runner/testResources/sample-project/MyMath.js': './packages/mocha-runner/testResources/sample-project-instrumented/MyMath.js',
+    './packages/mocha-runner/testResources/infinite-loop/infinite-loop.js': './packages/mocha-runner/testResources/infinite-loop-instrumented/infinite-loop.js',
   }, '__stryker2__');
   await instrument({
     './packages/jasmine-runner/testResources/jasmine-init/lib/jasmine_examples/Player.js': './packages/jasmine-runner/testResources/jasmine-init-instrumented/lib/jasmine_examples/Player.js',
-    './packages/jasmine-runner/testResources/jasmine-init/lib/jasmine_examples/Song.js': './packages/jasmine-runner/testResources/jasmine-init-instrumented/lib/jasmine_examples/Song.js'
+    './packages/jasmine-runner/testResources/jasmine-init/lib/jasmine_examples/Song.js': './packages/jasmine-runner/testResources/jasmine-init-instrumented/lib/jasmine_examples/Song.js',
+    './packages/jasmine-runner/testResources/infinite-loop/lib/infinite-loop.js': './packages/jasmine-runner/testResources/infinite-loop-instrumented/lib/infinite-loop.js',
   }, '__stryker2__');
   await instrument({
     './packages/karma-runner/testResources/sampleProject/src/Add.js': './packages/karma-runner/testResources/instrumented/src/Add.js',